requirements.txt
Flask==2.0.2
SQLAlchemy==1.4.25
PyJWT==1.5.3